--- Main/client/src/components/Footer.jsx
@@ -1,12 +1,0 @@
-// File: client/src/components/Footer.js
-import React from 'react';
-
-const Footer = () => {
-  return (
-    <footer>
-      <p>&copy; {new Date().getFullYear()} KIM Inventory Manager</p>
-    </footer>
-  );
-};
-
-export default Footer;
