+++ client/public/assets/android-chrome-512x512.png
