+++ client/public/assets/favicon.ico
