+++ client/src/components/Footer.jsx
@@ -1,9 +1,12 @@
+// File: client/src/components/Footer.js
+import React from 'react';
+
 const Footer = () => {
-    return (
-        <footer>
-            <p>&copy; 2024 KYM Inventory Management 💕</p>
-        </footer>
-    )
-}
+  return (
+    <footer>
+      <p>&copy; {new Date().getFullYear()} KIM Inventory Manager</p>
+    </footer>
+  );
+};
 
 export default Footer;
