+++ client/src/components/InventoryItemCard.jsx
