+++ client/src/pages/Inventory.jsx
