+++ models/InventoryItem.js
